+++ server/messenger/__init__.py
@@ -1,0 +1,3 @@
+from .wazzup import get_messenger, MessengerError, MessageData
+
+__all__ = ["get_messenger", "MessengerError", "MessageData"]